docs(api): 更新API文档增加积分抽奖消息通知和WebSocket接口
- 新增用户积分接口包括积分兑换商品功能
- 新增抽奖接口包括奖项配置更新和抽奖功能
- 新增消息通知接口支持商户发送消息给客户
- 新增WebSocket接口实现实时订单状态推送
- 更新订单接口增加积分相关字段和分页功能
- 更新商品接口增加积分兑换筛选功能
- 更新用户接口增加openId字段和签到状态
- 更新操作记录接口优化字段说明
- 优化API请求参数格式和响应示例
- 添加WebSocket连接和消息格式详细说明
- 更新文档时间戳至2026年3月31日

+++ src/main/java/cn/muzisheng/lebo/dto/PointRecordAddDTO.java
@@ -25,15 +25,15 @@ public class PointRecordAddDTO {
     /**
      * 变动积分数量
      */
-    private Integer changeAmount;
+    private Long changeAmount;
 
     /**
      * 变动前积分
      */
-    private Integer beforeAmount;
+    private Long beforeAmount;
 
     /**
      * 变动后积分
      */
-    private Integer afterAmount;
+    private Long afterAmount;
 }
